services/agent-runtime-py/pyproject.toml
[project]
name = "agent-runtime"
version = "0.1.0"
description = "Agent Runtime service — LangGraph executor agent (ReAct loop)"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "langchain-community>=0.3",
    "duckduckgo-search>=7.0",
    "langgraph-checkpoint-postgres>=2.0",
    "fpdf2>=2.8",
    "openpyxl>=3.1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
